Replace ForwardCancelable
